circulate-backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0

structlog==25.5.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg==3.2.9
psycopg-binary==3.2.9

celery==5.6.2
redis==5.0.8

canonicaljson==2.0.0

